Warnings Fix:
1. Removed unused variable in vpd_tool_impl.cpp.
2. Fix for sign compare warning in editor_impl.cpp & ibm_vpd_app.cpp.
3. Fix for unused variable warning in vpdecc files.

+++ vpd-manager/editor_impl.cpp
@@ -539,7 +539,7 @@ void EditorImpl::updateKeyword(const Binary& kwdData)
     vector<unsigned char>::const_iterator last = tmpVector.begin() + NEXT_64_KB;
 
     vpdFile = tmpVector;
-    if (distance(first, last) > tmpVector.size())
+    if (distance(first, last) > static_cast<int>(tmpVector.size()))
     {
         Binary extracted64KbVpd(first, last);
         vpdFile = extracted64KbVpd;
